Change write function return type to ssize_t for consistency
Updated the return type of the write function to ssize_t
to align with standard definitions and improve type safety.

+++ Tests/stdlib.c
@@ -5,7 +5,7 @@
 #include "rwonce.h"
 #include "tricore_testdevice.h"
 
-int write(int fd, const void *buf, size_t nbyte) {
+ssize_t write(int fd, const void *buf, size_t nbyte) {
   switch (fd) {
   case 1:
     /*
